+++ dexterity/membrane/behavior/membranepassword.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from password import IProvidePasswordsSchema as BBBIProvidePasswordsSchema
+
+
+class IProvidePasswordsSchema(BBBIProvidePasswordsSchema):
+    """ BBB class for zc.relations stuff and other weirdness """